Refactor: Update Firebase config and imports
Consolidate Firebase initialization and update imports to use the new alias.
