+++ mispronounciation/context/firebase.config.ts
@@ -1,5 +1,7 @@
 // firebase.config.ts
 import { initializeApp, getApps } from 'firebase/app';
+import { getAuth } from 'firebase/auth';
+import { getDatabase } from 'firebase/database';
 import { getFirestore, connectFirestoreEmulator } from 'firebase/firestore';
 import { getStorage, connectStorageEmulator } from 'firebase/storage';
 
@@ -7,6 +9,7 @@ const firebaseConfig = {
   // Replace with your actual Firebase configuration
   apiKey: "your-api-key-here",
   authDomain: "your-project.firebaseapp.com",
+  databaseURL: "https://your-project-id-default-rtdb.firebaseio.com",
   projectId: "your-project-id",
   storageBucket: "your-project.appspot.com",
   messagingSenderId: "123456789012",
@@ -16,18 +19,18 @@ const firebaseConfig = {
 // Initialize Firebase only if it hasn't been initialized already
 const app = getApps().length === 0 ? initializeApp(firebaseConfig) : getApps()[0];
 
-// Initialize Firestore
-const db = getFirestore(app);
-
-// Initialize Storage
+// Initialize Firebase services
+const auth = getAuth(app);
+const db = getDatabase(app); // Realtime Database
+const firestore = getFirestore(app); // Firestore
 const storage = getStorage(app);
 
 // For development/testing with Firebase emulators (optional)
 // Uncomment the lines below if you're using Firebase emulators
 /*
-if (__DEV__ && !db._delegate._databaseId.database.includes('(default)')) {
+if (__DEV__) {
   try {
-    connectFirestoreEmulator(db, 'localhost', 8080);
+    connectFirestoreEmulator(firestore, 'localhost', 8080);
     connectStorageEmulator(storage, 'localhost', 9199);
   } catch (error) {
     console.log('Emulator connection error:', error);
@@ -35,5 +38,5 @@ if (__DEV__ && !db._delegate._databaseId.database.includes('(default)')) {
 }
 */
 
-export { db, storage, app };
+export { auth, db, firestore, storage, app };
 export default app;
